switch to a more efficient bitarray implementation; todo: refactor driver.py, implement guided lookup
