+++ prototype/bitarray.py_bak
