+++ prototype/bloomfilter.py
@@ -1,4 +1,4 @@
-from bitarray import BitArray
+from bitarray import bitarray
 from fnv import hash_fnv
 from math import log, ceil
 
@@ -20,7 +20,8 @@ class BloomFilter:
             self.k = k
         self.fpp = fpp
         self.num_elements = n
-        self.ba = BitArray(num_bits)
+        self.ba = bitarray(num_bits)
+        self.ba.setall(False)
 
     def __str__(self):
         ''' Print some info for debugging.
@@ -31,52 +32,38 @@ class BloomFilter:
             BitArray size
             % of BitArray set
         '''
-        res = 'BloomFilter(fpp=%0.2f, n=%d, k=%d, ba=BitArray(%d, %%full=%0.1f))'\
+        res = 'BloomFilter(fpp=%0.2f, n=%d, k=%d, ba=(malloc=%dB, length=%db)'\
                 %(self.fpp,
                   self.num_elements,
                   self.k,
-                  self.ba.size,
-                  100.*self.ba.count_bits()/self.ba.size)
+                  self.ba.buffer_info()[-1],
+                  self.ba.length())
         return res
 
-    def _range(self, hashes=None, pattern=None):
-        '''Generate and return a list/generator of hash functions to use.
-        '''
-        if pattern is None:
-            if hashes is None:
-                return range(0, self.k, 1)
-            else:
-                return range(hashes[0], hashes[1]+1,1)
-        res = []
-        count = 0
-        while pattern:
-            if pattern & 1:
-                res.append(hashes[0]+count)
-            count += 1
-            pattern >>= 1
-        return res
+    def _set_bit(self, ix):
+        self.ba[ix] = True
+        return True
 
-    def insert(self, key, hashes=None, pattern=None):
+    def insert(self, key, hashes=[]):
         '''Insert key into Bloom filter. By default using full
-        range of hash functions. If hashes is a pair
-        of indices, will only use the range between hashes[0] and
-        hashes[1], inclusive.
+        range of hash functions. If hashes is a list/generator
+        of hash funcs to use (indices).
         '''
-        self._helper(key, lamda=self.ba.set_bit, 
-                     hashes=self._range(hashes, pattern),
+        self._helper(key, lamda=self._set_bit,
+                     hashes=hashes,
                      keep_going=True)
 
-    def contains(self, key, hashes=None, keep_going=False):
+    def contains(self, key, hashes=[], keep_going=False):
         '''Return 1 (possibly false positive) or 0.
         '''
-        return self._helper(key, lamda=self.ba.read_bit,
-                            hashes=self._range(hashes),
+        return self._helper(key, lamda=self.ba.__getitem__,
+                            hashes=hashes,
                             keep_going=keep_going)
 
 
     def _helper(self, key, lamda=None, hashes=[], keep_going=False):
         '''Insert or look up key (int) in Bloom filter using
-            hash functions from hashes list inclusive (by default).
+            hash functions from hashes list.
             If `keep_going` is True, complete all `hashes` and decode the result.
         '''
         if not hashes: return 0
@@ -84,12 +71,13 @@ class BloomFilter:
         h1 = hash64 & 0x00000000FFFFFFFF
         h2 = hash64 & 0xFFFFFFFF00000000
         decode = 0
+        size = self.ba.length()
         for i in hashes:
-            ix = (h1 + i * h2) % self.ba.size
+            ix = (h1 + i * h2) % size
             res = lamda(ix)
-            if res==0 and not keep_going:
+            if res==False and not keep_going:
                 return 0
-            decode += res<<i
+            decode += int(res)<<i
         return decode
 
 if __name__ == "__main__":
@@ -101,8 +89,8 @@ if __name__ == "__main__":
 
     # insert using only the first hash function
     for i in range(10):
-        bf.insert(i, hashes=(0,0))
+        bf.insert(i, hashes=[0])
 
     # look up only using the first hash function
     for i in range(15):
-        print(i, bf.contains(i, hashes=(0,0)))
+        print(i, bf.contains(i, hashes=[0]))
