+++ requirements.txt
@@ -1,3 +1,3 @@
 matplotlib>=2.1.2
 netaddr>=0.7.19
-
+bitarray>=0.8.1
